feat: support conversational runtime prompts
