+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "uipath-langchain"
-version = "0.16.1"
+version = "0.16.2"
 description = "Python SDK that enables developers to build and deploy LangGraph agents to the UiPath Cloud Platform"
 readme = { file = "README.md", content-type = "text/markdown" }
 requires-python = ">=3.11"
@@ -163,6 +163,7 @@ uipath = false
 uipath-core = false
 uipath-platform = false
 uipath-runtime = false
+uipath-dev = false
 uipath-langchain-client = false
 uipath-llm-client = false
 jsonschema-pydantic-converter = false
