+++ src/uipath_langchain/agent/advanced/types.py
@@ -4,7 +4,7 @@ from typing import Annotated, Any
 
 from langchain_core.messages import AnyMessage
 from langgraph.graph.message import add_messages
-from pydantic import BaseModel
+from pydantic import BaseModel, ConfigDict
 
 
 class AdvancedAgentGraphState(BaseModel):
@@ -14,8 +14,13 @@ class AdvancedAgentGraphState(BaseModel):
     structured_response: dict[str, Any] = {}
 
 
-class ConversationalAdvancedAgentGraphState(BaseModel):
-    """Graph state for the conversational advanced agent wrapper."""
+class _ConversationalAdvancedAgentGraphInput(BaseModel):
+    model_config = ConfigDict(validate_by_alias=True, validate_by_name=True)
 
     messages: Annotated[list[AnyMessage], add_messages] = []
+
+
+class ConversationalAdvancedAgentGraphState(_ConversationalAdvancedAgentGraphInput):
+    """Graph state for the conversational advanced agent wrapper."""
+
     initial_message_count: int | None = None
